DB型をfloat型へ変更
小数点以下の数値もDBに保存したい

+++ db/migrate/20210325014044_create_training_records.rb
@@ -3,7 +3,7 @@ class CreateTrainingRecords < ActiveRecord::Migration[6.0]
     create_table :training_records do |t|
       t.date       :date,            null: false
       t.integer    :training_event,  null: false
-      t.integer    :training_weight, null: false
+      t.float      :training_weight, null: false
       t.integer    :reps,            null: false
       t.integer    :set,             null: false 
       t.references :user,            foreign_key: true
